Renamed poisson_mixture.py to fit_poisson_mixture.py
